Register new users
Not certain about the functionality of `PDO::beginTransaction` &
`PDO::commit` to be sure that it will work when an existing user tries
to register or if the database crashes in the middle of the transaction.
Seems to at least not update database until `PDO::commit`.

Should probably look into improving `PDOStatament::execute`, so it
returns true|false, allowing me to detect any errors.