Check two forms of signature
Unfortunately, password reset signatures are screwed up by most email
clients. Copying and pasting the link results on one form of the
signature, and clicking on the link results in another. So, need to
check `urldecode`d version and then raw version.

Check  `urldecode` first, since that is the one more likely to verify
correctly.